feat(modules): rewrite cjsToEsm.cjs to cjsToEsm.mjs using ECMAScript modules
